Organize MediaFileDao
Processing remains the same ...

 - Change method order
 - Remove unused method
 - Remove MediaFileDao#getAlbumCount
   - Use countMediaFile
 - Remove getSongsCountForAlbum
   - Because the count value is already registered at the time of scanning

+++ jpsonic-main/src/main/java/com/tesshu/jpsonic/service/upnp/processor/RecentAlbumUpnpProcessor.java
@@ -65,7 +65,7 @@ public class RecentAlbumUpnpProcessor extends MediaFileUpnpProcessor implements 
 
     @Override
     public int getDirectChildrenCount() {
-        int count = mediaFileService.getAlbumCount(util.getGuestFolders());
+        int count = (int) mediaFileService.getAlbumCount(util.getGuestFolders());
         return Math.min(count, RECENT_COUNT);
     }
 }
